Added some plguins
Using neovim and configurin it is pretty hard. Specially with my limited experience in it. Yes I can navigate and edit code in neovim, but configuring plugins is hard. I plan to move in nvchad in the future because its much faster than lazyvim. But because of my limited knowledge of lua atm, configuring and writting code in lua just based on provided documentation is very slow, so I'll put it in the back burner for now nad leave it to my future self.